+++ global_sched.py
@@ -146,7 +146,7 @@ def push_task_into_bins(init_p_list: List[TaskInt], affinity,
                 _p.release_time += _p.task.period
                 _p.deadline += _p.task.period
 
-                _p.end_time = curr_t * timestep
+                _p.end_time = curr_t
                 _p.currentburst = 0
                 _p.burst = 0
                 _p.totburst = 0