+++ global_var.py
@@ -7,4 +7,5 @@ overhead_of_dequeuing_op = 1000 # cycles
 clock_period = 1e-9 # Seconds
 SRAM_size_per_core = 1.25 # MB
 
-
+trace_file = "trace.txt"
+trace_list = []