+++ msg_dispatcher.py
@@ -1,10 +1,14 @@
 # from multiprocessing import Process, Queue
 from queue import Queue
+from buffer import Data
 
 class MsgDispatcher:
-    def __init__(self, num_queues):
+    def __init__(self, num_queues, queue_list=None):
         self.num_processes = num_queues
-        self.queues = [Queue() for _ in range(num_queues)]
+        if queue_list is None:
+            self.queues = [Queue() for _ in range(num_queues)]
+        else:
+            self.queues = queue_list
 
     def broadcast_message(self, message, prefix=""):
         print(prefix+f"Broadcasting message: {message}")
@@ -22,6 +26,26 @@ class MsgDispatcher:
 #             process.start()
 #             processes.append(process)
 #         return processes
+
+def msg_read(msg_queue, curr_t, glb_name_p_dict, process_dict, buffer, bin_name, bin_event_flg):
+    msg_list = []
+    # read out all message and clear the message pipe
+    while not msg_queue.empty():
+        msg_list.append(msg_queue.get())
+    if msg_list:
+        for _p in process_dict.values():
+            for key, attr in _p.pred_data.items():
+                # if msg_pipe.filter(key):
+                if msg_filter(msg_list, key):
+                    attr["valid"] = True
+                    attr["time"] = curr_t
+                    # TODO: fix the event time as the actual time
+                    if bin_name and not bin_event_flg:
+                        bin_event_flg = True
+                        print(f"({bin_name})")
+                    print(f"		{_p.task.name} received event {key:s} @ {curr_t:.6f}")
+                    buffer.put(Data(glb_name_p_dict[key].pid, 1, (0,), "output", glb_name_p_dict[key].io_time, curr_t, 1/glb_name_p_dict[key].task.freq))
+    return bin_event_flg
 
 def msg_filter(msg_list:list, keyword:str):
     return [msg for msg in msg_list if keyword in msg]